+++ src/screens/Analytics/Components/CustomTooltip/CustomTooltip.scss
